Made a few changes and fixes
Changed was-blacklist in waiting data, fixed proposed-for-removal checking for miner in pending, added read-only functions to get remaining blocks until users in pending list can enter the pool, to get the status for the given address (also added a private function to check if the user is in pending list), and to get a list of miners proposed for removal. Changed a few public functions to private.